+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * ability to include more edge/node tags into the graph
+* automatically tag nodes part of a roundabout as junction=roundabout
+
 ### Changed
 
 * dropped unused oneway convertion